Update tests to reflect new reader/writer hierarchy
